USA dataset refactors

+++ tools/wonderTxt2Csv.sh
@@ -8,6 +8,9 @@ FILE_NEW=$(
 
 cp "$FILE" "$FILE_NEW.csv"
 
-sed -i -e 's/	/,/g' "$FILE_NEW.csv" # Replace Tabs with comma
+# Replace Tabs with comma, remove comments
+sed -i -e 's/	/,/g;/---/,$d' "$FILE_NEW.csv"
 
-rm "$FILE_NEW.csv-e"
+if [ "$(uname)" == "Darwin" ]; then
+  rm "$FILE_NEW.csv-e"
+fi
